Refactor and Reorganize (group) code a bit
