--- android/src/main/java/com/terrylinla/rnsketchcanvas/utils/BorderStyle.java
@@ -1,5 +1,0 @@
-package com.terrylinla.rnsketchcanvas.utils;
-
-public enum BorderStyle {
-    DASHED, SOLID
-}
