+++ android/src/main/java/com/terrylinla/rnsketchcanvas/utils/entities/BorderStyle.java
@@ -1,0 +1,5 @@
+package com.terrylinla.rnsketchcanvas.utils.entities;
+
+public enum BorderStyle {
+    DASHED, SOLID
+}
