+++ android/src/main/java/com/terrylinla/rnsketchcanvas/utils/layers/Layer.java
@@ -11,14 +11,17 @@ public class Layer {
     private float mRotationInDegrees;
 
     private float mScale;
+    
     /**
      * top left X coordinate, relative to parent canvas
      */
     private float x;
+    
     /**
      * top left Y coordinate, relative to parent canvas
      */
     private float y;
+    
     /**
      * is layer flipped horizontally (by X-coordinate)
      */
